Move method.js file into method directory
