+++ lib/plugins/aws/deploy/compile/events/apiGateway/index.js
@@ -7,7 +7,7 @@ const compileRestApi = require('./lib/restApi');
 const compileApiKeys = require('./lib/apiKeys');
 const compileResources = require('./lib/resources');
 const compileCors = require('./lib/cors');
-const compileMethods = require('./lib/methods');
+const compileMethods = require('./lib/method/index');
 const compileAuthorizers = require('./lib/authorizers');
 const compileDeployment = require('./lib/deployment');
 const compilePermissions = require('./lib/permissions');
