+++ lib/plugins/aws/deploy/compile/events/apiGateway/lib/method/index.js
